MatPlotLib Overhaul
Because it seemed that getting click input was going to be overcomplicated as hell, I decided to forgo Plotly in favor of MatPlotLib. This also allowed me to stop using the CEFPython module.

As of this update, many of the features are currently broken, but this overhaul should allow me to rebuild them to be more robust.

Old code is in the Deprecated folder.