Click Handling Update (Overhaul pt. II)
Part II of the overhaul. Some features are still missing, including plot exporting, zooming on the plot, and the peak table.

Click handling is live, so peaks can be picked by clicking the start and end points. Peak picking works fine; the data are just not displayed anywhere on the GUI.

Undo/redo functionality are up and running (with a few bugs).

Finally, I added a bunch of comments and docstrings, as well as just generally cleaning things up and making the code more PEP-8 compliant.

Check agenda.txt for planned features.